+++ _examples/handler.go
@@ -30,14 +30,14 @@ type myHandler struct{}
 // Customize your own handler.
 func (mh *myHandler) Handle(log *logit.Log) bool {
 	os.Stdout.Write([]byte("myHandler: "))
-	os.Stdout.Write(logit.EncoderOf("json").Encode(log, "")) // Try `os.Stdout.WriteString(log.Msg())` ?
+	os.Stdout.Write(logit.TextEncoder().Encode(log, "")) // Try `os.Stdout.WriteString(log.Msg())` ?
 	return true
 }
 
-// We recommend you to register your handler to logit, so that
-// you can use your handler by logit.HandlerOf.
-// See logit.HandlerOf.
 func init() {
+	// We recommend you to register your handler to logit, so that
+	// you can use your handler in config file.
+	// See logit.RegisterHandler.
 	logit.RegisterHandler("myHandler", func(params map[string]interface{}) logit.Handler {
 		return &myHandler{}
 	})
@@ -45,23 +45,22 @@ func init() {
 
 func main() {
 
-	// Create a logger holder.
-	// Default handler is logit.DefaultHandler.
-	logger := logit.NewDevelopLogger()
+	// Create a logger holder with a console handler.
+	logger := logit.NewLogger(logit.DebugLevel, logit.NewConsoleHandler(logit.TextEncoder(), logit.DefaultTimeFormat))
 	logger.Info("before adding handlers...")
 	fmt.Println("fmt =========================================")
 
 	// Add handlers to logger.
-	// There are three handlers in logger because logger has one handler inside after creating.
-	// See logit.ConsoleHandler.
-	logger.AddHandlers(&myHandler{}, logit.HandlerOf("console", map[string]interface{}{}))
-	logger.Info("after adding handlers...")
+	// There are three handlers in logger because logger has one handler inside before adding.
+	// See logit.NewConsoleHandler.
+	logger.AddHandlers(&myHandler{}, logit.NewConsoleHandler(logit.JsonEncoder(), ""))
+	logger.Info("after adding two handlers...")
 	fmt.Println("fmt =========================================")
 
 	// Set handlers to logger.
 	// There are one handler in logger because all handlers inside was removed.
 	// If you register your handler to logit by logit.RegisterHandler, then you can
 	// use your handler everywhere like this:
-	logger.SetHandlers(logit.HandlerOf("myHandler", map[string]interface{}{}))
-	logger.Info("after setting handlers...")
+	logger.SetHandlers(&myHandler{})
+	logger.Info("after setting one handlers...")
 }