+++ config.go
@@ -24,98 +24,76 @@ import (
 	"io/ioutil"
 )
 
-// Config is for configuring a Logger.
-// You can use a config to create a Logger to use.
-// See NewLoggerFrom(config Config).
-type Config struct {
+// config is the config mapping the config file.
+type config struct {
 
-	// Level is the level of Logger.
-	// If the level of log is smaller than this Level, this log will be ignored.
-	Level Level
-
-	// NeedCaller will determine weather you need caller info or not.
-	// Notice that adding caller will use runtime method which costs lots of time,
-	// so set it to true only when you really need it.
-	NeedCaller bool
-
-	// Handlers is how to handle a log in logger.
-	// We provide some handlers and you can use them directly.
-	// See logit.Handler.
-	Handlers []Handler
-}
-
-// fileConfig is the config mapping a file.
-type fileConfig struct {
-
-	// Level is the level in string form.
+	// Level is the logger level in string form.
 	Level string `json:"level"`
 
-	// Caller will determine weather you need caller info or not.
-	// Notice that adding caller will use runtime method which costs lots of time,
+	// Caller will determine weather you need caller info in log or not.
+	// Notice that adding caller will call a runtime method which costs lots of time,
 	// so set it to true only when you really need it.
 	Caller bool `json:"caller"`
 
-	// Handlers is the mapping to config file.
+	// Handlers is the handlers you want to use mapping to config file.
 	Handlers map[string]map[string]interface{} `json:"handlers"`
 }
 
-// removeComments removes all comments of fileInBytes.
-func removeComments(fileInBytes []byte) []byte {
-	// 注释只能是单独起一行，并且以 # 开头
+// jsonifyContent normalize content to a standard json string in bytes.
+// We make a new config format which is transformed from Json.
+// As we know, a standard Json string is strict in format. However,
+// the worst thing when using a Json as config file is we can't add any
+// comments so we have no idea that why we use this configuration in our config.
+// So we must transform it. The first thing we do is let it supports comments.
+// Also, the "{}" in Json string is redundant in a config file. After doing these,
+// a reading-friendly config file is born, and it is not a standard Json string any more.
+// This method jsonifyContent will fix this so that we can use Json parser to
+// parse our config file.
+func jsonifyContent(content []byte) []byte {
+
+	// 我们的配置文件是支持注释的，而 Json 规范中并没有对注释的支持，所以我们需要对注释进行擦除
+	// 注意：注释只能是单独起一行，并且以 # 开头
 	var buffer []byte
-	lines := bytes.Split(fileInBytes, []byte("\n"))
+	lines := bytes.Split(content, []byte("\n"))
 	for _, line := range lines {
 		if !bytes.HasPrefix(bytes.TrimSpace(line), []byte("#")) {
 			buffer = append(buffer, line...)
 		}
 	}
-	return buffer
-}
-
-// parseFileConfig parses fileConfig and convert it to Config.
-// Return an error if something wrong happened.
-func parseFileConfig(fileConfig fileConfig) (Config, error) {
-
-	// 解析日志级别
-	level, err := ParseLevel(fileConfig.Level)
-	if err != nil {
-		return Config{}, err
-	}
-
-	// 解析日志处理器
-	var handlers []Handler
-	for name, parmas := range fileConfig.Handlers {
-		handlers = append(handlers, HandlerOf(name, parmas))
-	}
-
-	return Config{
-		Level:      level,
-		NeedCaller: fileConfig.Caller,
-		Handlers:   handlers,
-	}, nil
-}
-
-// ParseConfigFile parses a config file whose path is configFile and returns
-// a Config of it. Return an error if something wrong happened.
-func ParseConfigFile(configFile string) (Config, error) {
-
-	// 配置文件一般不会太大，直接全部读取进内存
-	fileInBytes, err := ioutil.ReadFile(configFile)
-	if err != nil {
-		return Config{}, err
-	}
 
 	// 由于配置文件使用 Json 格式，而 Json 规范要求使用 {} 包裹内容，但这个 {} 不方便配置文件的阅读，
 	// 所以我们设定在配置文件中不使用 {} 包裹，而是交给我们读取出来之后进行包裹
-	// 另外，我们的配置文件是支持注释的，而 Json 规范中并没有对注释的支持，所以我们需要对注释进行擦除
-	configInBytes := bytes.Join([][]byte{[]byte("{"), removeComments(fileInBytes), []byte("}")}, []byte(""))
-	fileConfig := fileConfig{
-		Level: "debug",
-	}
-	err = json.Unmarshal(configInBytes, &fileConfig)
+	return bytes.Join([][]byte{[]byte("{"), buffer, []byte("}")}, []byte(""))
+}
+
+// parseConfigFile parses a config file whose path is configFile and returns
+// a mapping config of it. Return an error if something wrong happened.
+func parseConfigFile(configFile string) (config, error) {
+
+	// 配置文件一般不会太大，直接全部读取进内存
+	content, err := ioutil.ReadFile(configFile)
 	if err != nil {
-		return Config{}, err
+		return config{}, err
 	}
 
-	return parseFileConfig(fileConfig)
+	// 反序列化出配置
+	conf := config{
+		Level: "debug",
+	}
+	err = json.Unmarshal(jsonifyContent(content), &conf)
+	if err != nil {
+		return config{}, err
+	}
+
+	return conf, nil
+}
+
+// parseHandlersFromConfig parses all handlers in conf.
+// Return a slice of all parsed handlers.
+func parseHandlersFromConfig(conf config) []Handler {
+	handlers := make([]Handler, 0, len(conf.Handlers)+2)
+	for name, params := range conf.Handlers {
+		handlers = append(handlers, handlerOf(name, params))
+	}
+	return handlers
 }