+++ encoder.go
@@ -20,22 +20,20 @@ package logit
 
 import (
 	"bytes"
-	"errors"
+	"fmt"
+	"os"
 	"strconv"
 	"strings"
 )
 
 var (
-	// encoders stores all encoders provided.
-	// Call EncoderOf method to use one of encoders below.
+	// encoders store all encoders provided.
+	// Call encoderOf method to use one of encoders below.
 	// Actually, this field is for me, not for you, ha:)
 	encoders = map[string]Encoder{
-		"text": EncodeToText,
-		"json": EncodeToJson,
+		"text": TextEncoder(),
+		"json": JsonEncoder(),
 	}
-
-	// EncoderIsNotExistedError is the error which will occur on pointing a nonexistent encoder.
-	EncoderIsNotExistedError = errors.New("the encoder you pointed is not existed! Try text or json")
 )
 
 // Encoder is for encoding a log to bytes with timeFormat.
@@ -48,76 +46,87 @@ func (e Encoder) Encode(log *Log, timeFormat string) []byte {
 	return e(log, timeFormat)
 }
 
-// EncoderOf returns the encoder called name.
-// Notice that it will panic if the encoder called name doesn't exist.
-func EncoderOf(name string) Encoder {
+// encoderOf returns the encoder called name.
+// If the encoder doesn't exist, a tip will be printed and
+// the program will exit with status code -2.
+func encoderOf(name string) Encoder {
 	encoder, ok := encoders[name]
 	if !ok {
-		panic(EncoderIsNotExistedError)
+		fmt.Fprintf(os.Stderr, "Error: The encoder \"%s\" you pointed doesn't exist! Try \"text\" or \"json\".\n", name)
+		os.Exit(-2)
 	}
 	return encoder
 }
 
-// EncodeToText encodes a log to a plain string like "[Info] [2020-03-06 16:10:44] msg" in bytes.
-func EncodeToText(log *Log, timeFormat string) []byte {
+// =================================== text encoder ===================================
 
-	// 组装 log
-	buffer := bytes.NewBuffer(make([]byte, 0, 64))
-	buffer.WriteString("[")
-	buffer.WriteString(log.Level().String())
-	buffer.WriteString("] [")
-
-	// 判断是否需要格式化时间
-	if timeFormat != "" {
-		buffer.WriteString(log.Now().Format(timeFormat))
-	} else {
-		buffer.WriteString(strconv.FormatInt(log.Now().Unix(), 10))
-	}
-
-	buffer.WriteString("] ")
-
-	// 如果有文件信息，就把文件信息也加进去
-	if log.file != "" && log.Line() != 0 {
-		buffer.WriteString("[")
-		buffer.WriteString(log.File() + ":" + strconv.Itoa(log.Line()))
-		buffer.WriteString("] ")
-	}
-
-	buffer.WriteString(log.Msg())
-	buffer.WriteString("\n")
-	return buffer.Bytes()
-}
-
-// EncodeToJson encodes a log to a Json string like `{"level":"debug", "time":"2020-03-22 22:35:00", "msg":"log content..."}` in bytes.
+// TextEncoder encodes a log to a plain string like "[Info] [2020-03-06 16:10:44] msg" in bytes.
 // If timeFormat == "", then it will not format time and keep time in unix form.
-func EncodeToJson(log *Log, timeFormat string) []byte {
+func TextEncoder() Encoder {
+	return func(log *Log, timeFormat string) []byte {
 
-	// 组装 log
-	buffer := bytes.NewBuffer(make([]byte, 0, 64))
-	buffer.WriteString(`{"level":"`)
-	buffer.WriteString(log.Level().String())
-	buffer.WriteString(`", "time":`)
+		// 组装 log
+		buffer := bytes.NewBuffer(make([]byte, 0, 64))
+		buffer.WriteString("[")
+		buffer.WriteString(log.Level().String())
+		buffer.WriteString("] [")
 
-	// 判断是否需要格式化时间
-	if timeFormat != "" {
-		buffer.WriteString(strconv.Quote(log.Now().Format(timeFormat)))
-	} else {
-		buffer.WriteString(strconv.FormatInt(log.Now().Unix(), 10))
+		// 判断是否需要格式化时间
+		if timeFormat != "" {
+			buffer.WriteString(log.Now().Format(timeFormat))
+		} else {
+			buffer.WriteString(strconv.FormatInt(log.Now().Unix(), 10))
+		}
+
+		buffer.WriteString("] ")
+
+		// 如果有文件信息，就把文件信息也加进去
+		if log.file != "" && log.Line() != 0 {
+			buffer.WriteString("[")
+			buffer.WriteString(log.File() + ":" + strconv.Itoa(log.Line()))
+			buffer.WriteString("] ")
+		}
+
+		buffer.WriteString(log.Msg())
+		buffer.WriteString("\n")
+		return buffer.Bytes()
 	}
-
-	// 如果有文件信息，就把文件信息也加进去
-	if log.file != "" && log.Line() != 0 {
-		buffer.WriteString(`, "file":"` + log.File())
-		buffer.WriteString(`", "line":` + strconv.Itoa(log.Line()))
-	}
-
-	buffer.WriteString(`, "msg":"`)
-	buffer.WriteString(escapeString(log.Msg()))
-	buffer.WriteString("\"}\n")
-	return buffer.Bytes()
 }
 
-// escapeString is for escaping string from special character, such as double quotes.
+// =================================== json encoder ===================================
+
+// JsonEncoder encodes a log to a Json string like `{"level":"debug", "time":"2020-03-22 22:35:00", "msg":"log content..."}` in bytes.
+// If timeFormat == "", then it will not format time and keep time in unix form.
+func JsonEncoder() Encoder {
+	return func(log *Log, timeFormat string) []byte {
+
+		// 组装 log
+		buffer := bytes.NewBuffer(make([]byte, 0, 64))
+		buffer.WriteString(`{"level":"`)
+		buffer.WriteString(log.Level().String())
+		buffer.WriteString(`", "time":`)
+
+		// 判断是否需要格式化时间
+		if timeFormat != "" {
+			buffer.WriteString(strconv.Quote(log.Now().Format(timeFormat)))
+		} else {
+			buffer.WriteString(strconv.FormatInt(log.Now().Unix(), 10))
+		}
+
+		// 如果有文件信息，就把文件信息也加进去
+		if log.file != "" && log.Line() != 0 {
+			buffer.WriteString(`, "file":"` + log.File())
+			buffer.WriteString(`", "line":` + strconv.Itoa(log.Line()))
+		}
+
+		buffer.WriteString(`, "msg":"`)
+		buffer.WriteString(escapeString(log.Msg()))
+		buffer.WriteString("\"}\n")
+		return buffer.Bytes()
+	}
+}
+
+// escapeString is for escaping string from special characters, such as double quotes.
 // See issue: https://github.com/FishGoddess/logit/issues/1
 func escapeString(s string) string {
 
@@ -131,7 +140,7 @@ func escapeString(s string) string {
 			builder.WriteRune('\\')
 			builder.WriteRune(r)
 		default:
-			// ascii 小于 16 需要在前面补 \u000，介于 16 和 32 之间的需要补 \u00
+			// ascii 小于 16 的需要在前面补 \u000，介于 [16, 32) 之间的需要补 \u00
 			if r < 16 {
 				builder.WriteString("\\u000" + strconv.FormatInt(int64(r), 16))
 			} else if r < 32 {