+++ level.go
@@ -19,11 +19,12 @@
 package logit
 
 import (
-	"errors"
+	"fmt"
 	"math"
+	"os"
 )
 
-// Level is the type representation of the level.
+// Level is the type representation of the logger level.
 type Level uint8
 
 const (
@@ -37,7 +38,7 @@ const (
 )
 
 var (
-	// levels provides the name of all supported level.
+	// levels store the names of all level provided.
 	levels = map[Level]string{
 		DebugLevel: "debug",
 		InfoLevel:  "info",
@@ -45,24 +46,24 @@ var (
 		ErrorLevel: "error",
 		OffLevel:   "off",
 	}
-
-	// LevelIsNotExisted is an error representation of a level is not existed.
-	LevelIsNotExisted = errors.New("level is not existed, be sure your level is one of them: debug, info, warn, error, off")
 )
 
-// ParseLevel parses level and returns the Level of it.
-// Return LevelIsNotExisted if level is not existed.
-func ParseLevel(level string) (Level, error) {
+// parseLevel parses level and returns the Level of it.
+// If the level doesn't exist, a tip will be printed and
+// the program will exit with status code -3.
+func parseLevel(level string) Level {
 	for k, v := range levels {
 		if v == level {
-			return k, nil
+			return k
 		}
 	}
-	return 0, LevelIsNotExisted
+	fmt.Fprintf(os.Stderr, "Error: Level \"%s\" doesn't exist! Be sure your level is one of them: debug, info, warn, error, off\n", level)
+	os.Exit(-3)
+	return OffLevel
 }
 
-// The String method is used to print values passed as an operand
-// to any format that accepts a string or to an printer without format such as Print.
+// String returns the name of Level ll.
+// This method will be called when using printing operations like fmt.Println.
 func (ll Level) String() string {
 	return levels[ll]
 }