+++ logit.go
@@ -19,8 +19,8 @@
 package logit
 
 // globalLogger is a Logger holder for global usage.
-// Default level is info level.
-var globalLogger = NewDevelopLogger()
+// Default level is debug level.
+var globalLogger = NewLogger(DebugLevel, NewConsoleHandler(TextEncoder(), DefaultTimeFormat))
 
 // Me returns globalLogger for more usages.
 func Me() *Logger {
@@ -86,28 +86,28 @@ func Error(msg string) {
 
 // DebugFunc will output msg as a debug message.
 // The msg is the return value of msgGenerator.
-// This is a better way to output a long log made of many variables.
+// This is the better way to output a long log made from many variables.
 func DebugFunc(msgGenerator func() string) {
 	globalLogger.log(callDepthOfGlobalLogger, DebugLevel, msgGenerator())
 }
 
 // InfoFunc will output msg as an info message.
 // The msg is the return value of msgGenerator.
-// This is a better way to output a long log made of many variables.
+// This is the better way to output a long log made from many variables.
 func InfoFunc(msgGenerator func() string) {
 	globalLogger.log(callDepthOfGlobalLogger, InfoLevel, msgGenerator())
 }
 
 // WarnFunc will output msg as a warn message.
 // The msg is the return value of msgGenerator.
-// This is a better way to output a long log made of many variables.
+// This is the better way to output a long log made from many variables.
 func WarnFunc(msgGenerator func() string) {
 	globalLogger.log(callDepthOfGlobalLogger, WarnLevel, msgGenerator())
 }
 
 // ErrorFunc will output msg as an error message.
 // The msg is the return value of msgGenerator.
-// This is a better way to output a long log made of many variables.
+// This is the better way to output a long log made from many variables.
 func ErrorFunc(msgGenerator func() string) {
 	globalLogger.log(callDepthOfGlobalLogger, ErrorLevel, msgGenerator())
 }